requirements.txt
streamlit
pandas
matplotlib
groq
python-dotenv